+++ apps/api/urls/connect_urls.py
@@ -4,7 +4,7 @@ from ..views import connect_views as views
 
 app_name = 'Connect endpoints'
 urlpatterns = [
-    path('create/<int:pk>', views.create_connect),
+    path('create/', views.create_connect),
     path('read/<int:pk>', views.read_connect),
     path('read/list/<int:pk>', views.read_connect_list),
     path('update/<int:pk>', views.update_connect),